Require tlog_id when log_id_ranges is passed in
tlog_id specifes the active shard and is kept for backwards compatibility.
To avoid replicating information, the log_id_ranges flag is now used only to
specify inactive shards and must be used in conjunction with a tlog_id flag.
The log_id_ranges flag together with the tlog_id flag build a logRanges type,
which is wrapped by a logRangesFlag.

Signed-off-by: Lily Sturmann <lsturman@redhat.com>